AppForenseAntiMalware/edr/requirements.txt
requests==2.31.0
psutil==5.9.6
watchdog==3.0.0
yara-python==4.5.0
pydantic==2.5.2
python-dotenv==1.0.0
